+++ native_library/source/swift/CoreMlController.swift
@@ -17,11 +17,22 @@ import Foundation
 import FreSwift
 import CoreML
 import Vision
+#if os(iOS)
+import AVFoundation
+#endif
 class CoreMlController: NSObject, FreSwiftController {
     var TAG: String? = "CoreMlController"
     var context: FreContextSwift!
-    private var models: [String: MLModel] = [:]
-    private let backgroundQueue = DispatchQueue(label: "com.tuarua.mlane.backgroundQueue", qos: .background)
+    internal var models: [String: MLModel] = [:]
+    internal let userInitiatedQueue = DispatchQueue(label: "com.tuarua.mlane.userInitiatedQueue", qos: .userInitiated)
+#if os(iOS)
+    internal var captureSession: AVCaptureSession?
+    internal var videoPreviewLayer: AVCaptureVideoPreviewLayer?
+    internal var cameraView: UIView?
+    internal var visionModel: VNCoreMLModel?
+    internal var visionModelId: String?
+    internal var visionModelResult: [String: Any] = Dictionary()
+#endif
     convenience init(context: FreContextSwift) {
         self.init()
         self.context = context
@@ -32,11 +43,10 @@ class CoreMlController: NSObject, FreSwiftController {
         props["id"] = id
         guard let modelUrl = URL(safe: path) else {
             props["error"] = "invalid path"
-            let json = JSON(props)
-            self.sendEvent(name: ModelEvent.ERROR, value: json.description)
+            self.sendEvent(name: ModelEvent.ERROR, value: JSON(props).description)
             return
         }
-        backgroundQueue.async {
+        userInitiatedQueue.async {
             do {
                 let tmpUrl = try MLModel.compileModel(at: modelUrl)
                 let fileManager = FileManager.default
@@ -51,17 +61,14 @@ class CoreMlController: NSObject, FreSwiftController {
                         try fileManager.copyItem(at: tmpUrl, to: compiledUrl)
                     }
                     props["path"] = compiledUrl.absoluteString
-                    let json = JSON(props)
-                    self.sendEvent(name: ModelEvent.COMPILED, value: json.description)
+                    self.sendEvent(name: ModelEvent.COMPILED, value: JSON(props).description)
                 } catch {
                     props["error"] = error.localizedDescription
-                    let json = JSON(props)
-                    self.sendEvent(name: ModelEvent.ERROR, value: json.description)
+                    self.sendEvent(name: ModelEvent.ERROR, value: JSON(props).description)
                 }
             } catch let error {
                 props["error"] = error.localizedDescription
-                let json = JSON(props)
-                self.sendEvent(name: ModelEvent.ERROR, value: json.description)
+                self.sendEvent(name: ModelEvent.ERROR, value: JSON(props).description)
             }
         }
     }
@@ -69,72 +76,69 @@ class CoreMlController: NSObject, FreSwiftController {
     func loadModel(id: String, path: String) {
         var props: [String: Any] = Dictionary()
         props["id"] = id
-        backgroundQueue.async {
+        userInitiatedQueue.async {
             do {
                 guard let url = URL(safe: path) else {
                     props["error"] = "invalid path"
-                    let json = JSON(props)
-                    self.sendEvent(name: ModelEvent.ERROR, value: json.description)
+                    self.sendEvent(name: ModelEvent.ERROR, value: JSON(props).description)
                     return
                 }
                 self.models[id] = try MLModel(contentsOf: url)
                 props["path"] = path
-                let json = JSON(props)
-                self.sendEvent(name: ModelEvent.LOADED, value: json.description)
+                self.sendEvent(name: ModelEvent.LOADED, value: JSON(props).description)
             } catch let error {
                 props["error"] = error.localizedDescription
-                let json = JSON(props)
-                self.sendEvent(name: ModelEvent.ERROR, value: json.description)
+                self.sendEvent(name: ModelEvent.ERROR, value: JSON(props).description)
             }
         }
     }
     
+    func disposeModel(id: String) {
+        models[id] = nil
+    }
+    
     func prediction(id: String, input: [String: MLFeatureValue], maxResults: Int) {
-        if let model = models[id] {
-            var props: [String: Any] = Dictionary()
-            props["id"] = id
-            let modelDescription = model.modelDescription
-            let featureProvider = BaseInput.init(modelDescription: modelDescription)
-            featureProvider.setValues(dictionary: input, context)
-            DispatchQueue.main.async {
-                do {
-                    let prediction = try model.prediction(from: featureProvider)
-                    for featureName in prediction.featureNames {
-                        var feature: [String: Any] = Dictionary()
-                        if let val = prediction.featureValue(for: featureName) {
-                            switch val.type {
-                            case .dictionary:
-                                let dictionaryValue = val.dictionaryValue
-                                if dictionaryValue.isEmpty {
-                                    feature["dictionaryV"] = [:]
-                                } else {
-                                    let slicedArray = dictionaryValue.sorted { $0.value > $1.value }.prefix(maxResults)
-                                    feature["dictionaryV"] = slicedArray.map { item in
-                                        return ["k": item.key, "v": item.value]
-                                    }
-                                }
-                            case .double:
-                                feature["doubleV"] = val.doubleValue
-                            case .string:
-                                feature["stringV"] = val.stringValue
-                            case .int64:
-                                feature["int64V"] = val.int64Value
-                            default:
-                                break
+        guard let model = models[id] else { return  }
+
+        var props: [String: Any] = Dictionary()
+        props["id"] = id
+        let modelDescription = model.modelDescription
+        let featureProvider = BaseInput.init(modelDescription: modelDescription)
+        featureProvider.setValues(dictionary: input)
+        do {
+            let prediction = try model.prediction(from: featureProvider)
+            for featureName in prediction.featureNames {
+                var feature: [String: Any] = Dictionary()
+                if let val = prediction.featureValue(for: featureName) {
+                    switch val.type {
+                    case .dictionary:
+                        let dictionaryValue = val.dictionaryValue
+                        if dictionaryValue.isEmpty {
+                            feature["dictionaryV"] = [:]
+                        } else {
+                            let slicedArray = dictionaryValue.sorted { $0.value > $1.value }.prefix(maxResults)
+                            feature["dictionaryV"] = slicedArray.map { item in
+                                return ["k": item.key, "v": item.value]
                             }
-                            // TODO val.imageBufferValue
-                            // TODO val.multiArrayValue
-                            props[featureName] = feature
                         }
+                    case .double:
+                        feature["doubleV"] = val.doubleValue
+                    case .string:
+                        feature["stringV"] = val.stringValue
+                    case .int64:
+                        feature["int64V"] = val.int64Value
+                    default:
+                        break
                     }
-                    let json = JSON(props)
-                    self.sendEvent(name: ModelEvent.RESULT, value: json.description)
-                } catch let error {
-                    props["error"] = error.localizedDescription
-                    let json = JSON(props)
-                    self.sendEvent(name: ModelEvent.ERROR, value: json.description)
+                    // TODO val.imageBufferValue
+                    // TODO val.multiArrayValue
+                    props[featureName] = feature
                 }
             }
+            self.sendEvent(name: ModelEvent.RESULT, value: JSON(props).description)
+        } catch let error {
+            props["error"] = error.localizedDescription
+            self.sendEvent(name: ModelEvent.ERROR, value: JSON(props).description)
         }
     }