+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.0.3
+- iOS: camera input
+- iOS: HotDog Not HotDog example
+
 ### 0.0.2
 - allow multiple types of Models with different inputs/outputs
 - iOS version